requirements.txt
flask
azure-storage-blob==12.26.0